Chore #5 #6:  사소한 수정사항
- Book에 id속성 추가
- UserBookDetail의 process를 Float타입으로 변경
- UIImage 확장에 하트 아이콘 타입변수 추가

+++ Booky/Booky/Models/UserBookDetail.swift
@@ -16,8 +16,8 @@ struct UserBookDetail {
     private(set) var isLiked: Bool = false
     private(set) var totalPages: Int
     private(set) var readPages: Int = 0
-    var process: Double {
-        Double(readPages) / Double(totalPages)
+    var process: Float {
+        Float(readPages) / Float(totalPages)
     }
     
     //MARK: - Life Cycle
